@@ -4,10 +4,12 @@
 	"type": "module",
 	"private": true,
 	"scripts": {
-		"dev": "vite dev",
-		"dev:host": "vite dev --host",
-		"build": "vite build",
+		"prebuild-local-packages": "npm run prepare -w codemirror-lang-lxlquery && npm run build -w supersearch",
+		"dev": "npm run prebuild-local-packages && vite dev",
+		"dev:host": "npm run prebuild-local-packages && vite dev --host",
+		"build": "npm run prebuild-local-packages && vite build",
 		"preview": "vite preview",
+		"prepare": "svelte-kit sync || echo ''",
 		"test": "npm run test:integration && npm run test:unit",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
@@ -17,47 +19,46 @@
 		"test:unit": "vitest"
 	},
 	"devDependencies": {
-		"@axe-core/playwright": "^4.10.1",
-		"@eslint/compat": "^1.2.9",
-		"@floating-ui/dom": "^1.6.13",
-		"@iconify-json/bi": "^1.2.2",
-		"@playwright/test": "^1.51.0",
-		"@sveltejs/adapter-node": "^5.2.12",
-		"@sveltejs/kit": "^2.20.8",
-		"@sveltejs/vite-plugin-svelte": "^5.0.3",
-		"@tailwindcss/vite": "^4.1.3",
+		"@axe-core/playwright": "^4.10.2",
+		"@eslint/compat": "^1.3.2",
+		"@floating-ui/dom": "^1.7.3",
+		"@iconify-json/bi": "^1.2.6",
+		"@playwright/test": "^1.54.2",
+		"@sveltejs/adapter-node": "^5.2.14",
+		"@sveltejs/kit": "2.27.1",
+		"@sveltejs/vite-plugin-svelte": "^6.1.2",
+		"@tailwindcss/vite": "^4.1.11",
 		"@types/eslint": "^9.6.0",
 		"@types/jmespath": "^0.15.2",
 		"@types/js-cookie": "^3.0.6",
 		"codemirror-lang-lxlquery": "^0.1.0",
-		"cssnano": "^7.0.5",
 		"eslint": "^9.22.0",
-		"eslint-config-prettier": "^10.1.1",
-		"eslint-plugin-svelte": "^3.5.1",
+		"eslint-config-prettier": "^10.1.8",
+		"eslint-plugin-svelte": "^3.11.0",
 		"estree-walker": "^3.0.3",
-		"globals": "^16.0.0",
+		"globals": "^16.3.0",
 		"jmespath": "^0.16.0",
 		"js-cookie": "^3.0.5",
 		"lxljs": "^1.1.0",
 		"magic-string": "^0.30.11",
-		"mdsvex": "^0.12.3",
+		"mdsvex": "^0.12.6",
 		"nprogress": "^0.2.0",
-		"prettier": "^3.5.3",
-		"prettier-plugin-svelte": "^3.3.3",
-		"prettier-plugin-tailwindcss": "^0.6.11",
+		"prettier": "^3.6.2",
+		"prettier-plugin-svelte": "^3.4.0",
+		"prettier-plugin-tailwindcss": "^0.6.14",
 		"supersearch": "^0.0.1",
-		"svelte": "^5.28.2",
-		"svelte-check": "^4.1.7",
-		"tailwindcss": "^4.1.3",
+		"svelte": "^5.38.1",
+		"svelte-check": "^4.3.1",
+		"tailwindcss": "^4.1.11",
 		"tslib": "^2.6.3",
-		"typescript": "^5.8.2",
-		"typescript-eslint": "^8.26.1",
-		"unplugin-icons": "^22.1.0",
+		"typescript": "^5.9.2",
+		"typescript-eslint": "^8.39.1",
+		"unplugin-icons": "^22.2.0",
 		"vanilla-cookieconsent": "^3.1.0",
-		"vite": "^6.2.7",
-		"vitest": "^3.0.8"
+		"vite": "^7.1.2",
+		"vitest": "^3.2.4"
 	},
 	"dependencies": {
-		"dotenv": "^16.4.7"
+		"dotenv": "^17.2.1"
 	}
 }
