@@ -4,10 +4,10 @@
 	"description": "Libris XL query language support for CodeMirror",
 	"scripts": {
 		"test": "vitest",
-		"prepare": "rollup -c",
+		"prepare": "rollup -c --bundleConfigAsCjs --configImportAttributesKey with",
 		"lint": "eslint . && prettier --check .",
 		"format": "prettier --write .",
-		"watch": "rollup -c --watch"
+		"watch": "rollup -c --watch --bundleConfigAsCjs --configImportAttributesKey with"
 	},
 	"type": "module",
 	"main": "dist/index.cjs",
@@ -19,22 +19,22 @@
 	"types": "dist/index.d.ts",
 	"sideEffects": false,
 	"dependencies": {
-		"@codemirror/language": "^6.11.0",
+		"@codemirror/language": "^6.11.2",
 		"@lezer/highlight": "^1.0.0",
 		"@lezer/lr": "^1.0.0"
 	},
 	"devDependencies": {
-		"@lezer/generator": "^1.7.2",
+		"@lezer/generator": "^1.8.0",
+		"@rollup/plugin-typescript": "^12.1.4",
 		"eslint": "^9.22.0",
-		"eslint-config-prettier": "^10.1.1",
-		"globals": "^16.0.0",
-		"prettier": "^3.5.3",
-		"rollup": "^2.79.2",
-		"rollup-plugin-dts": "^4.0.1",
-		"rollup-plugin-ts": "^3.0.2",
-		"typescript": "^4.9.5",
-		"typescript-eslint": "^8.26.1",
-		"vitest": "^3.0.8"
+		"eslint-config-prettier": "^10.1.8",
+		"globals": "^16.3.0",
+		"prettier": "^3.6.2",
+		"rollup": "^4.46.2",
+		"rollup-plugin-dts": "^6.2.1",
+		"typescript": "^5.9.2",
+		"typescript-eslint": "^8.39.1",
+		"vitest": "^3.2.4"
 	},
 	"license": "MIT"
 }
