@@ -3,18 +3,17 @@
 	"version": "0.0.1",
 	"scripts": {
 		"dev": "vite dev",
-		"build": "vite build && npm run package",
+		"build": "vite build && npm run prepack",
 		"preview": "vite preview",
-		"package": "svelte-kit sync && svelte-package && publint",
-		"prepublishOnly": "npm run package",
+		"prepare": "svelte-kit sync || echo ''",
+		"prepack": "svelte-kit sync && svelte-package && publint",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"test:e2e": "playwright test",
-		"test": "npm run test:e2e && npm run test:unit -- --run",
-		"test:unit": "vitest",
-		"lint": "eslint . && prettier --check .",
 		"format": "prettier --write .",
-		"prepare": "npm run package",
+		"lint": "prettier --check . && eslint .",
+		"test:unit": "vitest",
+		"test": "npm run test:unit -- --run && npm run test:e2e",
+		"test:e2e": "playwright test",
 		"watch": "svelte-kit sync && svelte-package -w"
 	},
 	"files": [
@@ -38,32 +37,32 @@
 		"svelte": "^5.0.0"
 	},
 	"devDependencies": {
-		"@codemirror/commands": "^6.8.0",
+		"@codemirror/commands": "^6.8.1",
 		"@codemirror/state": "^6.5.2",
-		"@codemirror/view": "^6.36.4",
-		"@eslint/compat": "^1.2.9",
-		"@playwright/test": "^1.51.0",
-		"@sveltejs/adapter-auto": "^6.0.0",
-		"@sveltejs/kit": "^2.20.8",
-		"@sveltejs/package": "^2.3.11",
-		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@codemirror/view": "6.36.6",
+		"@eslint/compat": "^1.3.2",
+		"@playwright/test": "^1.54.2",
+		"@sveltejs/adapter-auto": "^6.1.0",
+		"@sveltejs/kit": "2.27.1",
+		"@sveltejs/package": "^2.4.1",
+		"@sveltejs/vite-plugin-svelte": "^6.1.2",
 		"@types/eslint": "^9.6.0",
-		"codemirror": "^6.0.1",
+		"codemirror": "^6.0.2",
 		"codemirror-lang-lxlquery": "^0.1.0",
 		"eslint": "^9.22.0",
-		"eslint-config-prettier": "^10.1.1",
-		"eslint-plugin-svelte": "^3.5.1",
+		"eslint-config-prettier": "^10.1.8",
+		"eslint-plugin-svelte": "^3.11.0",
 		"esm-env": "^1.2.1",
-		"globals": "^16.0.0",
-		"prettier": "^3.5.3",
+		"globals": "^16.3.0",
+		"prettier": "^3.6.2",
 		"prettier-plugin-css-order": "^2.1.2",
-		"prettier-plugin-svelte": "^3.3.3",
-		"publint": "^0.2.0",
-		"svelte": "^5.28.2",
-		"svelte-check": "^4.1.7",
-		"typescript": "^5.8.2",
-		"typescript-eslint": "^8.26.1",
-		"vite": "^6.2.7",
-		"vitest": "^3.0.8"
+		"prettier-plugin-svelte": "^3.4.0",
+		"publint": "^0.3.12",
+		"svelte": "^5.38.1",
+		"svelte-check": "^4.3.1",
+		"typescript": "^5.9.2",
+		"typescript-eslint": "^8.39.1",
+		"vite": "^7.1.2",
+		"vitest": "^3.2.4"
 	}
 }
