@@ -2,8 +2,17 @@
 	import { env } from '$env/dynamic/public';
 	import { afterNavigate } from '$app/navigation';
 	import { setMatomoTracker, getMatomoTracker, setMatomoContext } from '$lib/contexts/matomo';
+	import { onMount, type Snippet } from 'svelte';
+	import { browser } from '$app/environment';
+
+	interface Props {
+		children?: Snippet;
+	}
+
+	let { children }: Props = $props();
 
 	const URL: string = env.PUBLIC_MATOMO_URL;
+	let scriptTag: HTMLScriptElement | undefined = $state();
 
 	setMatomoContext();
 	const tracker = getMatomoTracker();
@@ -28,11 +37,20 @@
 			}
 		}
 	});
+
+	onMount(() => {
+		if (browser) {
+			scriptTag?.addEventListener('load', onMatomoScriptLoad);
+		}
+		return () => {
+			scriptTag?.removeEventListener('load', onMatomoScriptLoad);
+		};
+	});
 </script>
 
 <svelte:head>
 	{#if URL}
-		<script async defer src={`${URL}/matomo.js`} on:load={onMatomoScriptLoad}></script>
+		<script bind:this={scriptTag} async defer src={`${URL}/matomo.js`}></script>
 	{/if}
 </svelte:head>
-<slot />
+{@render children?.()}
