@@ -1,21 +1,17 @@
 <script lang="ts">
-	import { page } from '$app/stores';
+	import { page } from '$app/state';
 	import { afterNavigate } from '$app/navigation';
 	import type { FacetSearch } from '$lib/types/search';
 	import BiSearch from '~icons/bi/search';
 
-	export let search: FacetSearch;
+	interface Props {
+		search: FacetSearch;
+	}
 
-	$: searchParams = new URLSearchParams(search.template.replace('/find?', ''));
-	$: qOutput = searchParams
-		.get('_q')
-		?.replace(
-			`{?${search.mapping.variable}}`,
-			`${rangeFrom ? `${search.mapping.variable}>=${rangeFrom}` : ''} ${rangeTo ? `${search.mapping.variable}<=${rangeTo}` : ''}`
-		);
+	let { search }: Props = $props();
 
-	let rangeFrom = search.mapping.greaterThanOrEquals;
-	let rangeTo = search.mapping.lessThanOrEquals;
+	let rangeFrom = $state(search.mapping.greaterThanOrEquals);
+	let rangeTo = $state(search.mapping.lessThanOrEquals);
 
 	afterNavigate(() => {
 		rangeFrom = search.mapping.greaterThanOrEquals;
@@ -27,15 +23,24 @@
 			e.preventDefault();
 		}
 	}
+	let searchParams = $derived(new URLSearchParams(search.template.replace('/find?', '')));
+	let qOutput = $derived(
+		searchParams
+			.get('_q')
+			?.replace(
+				`{?${search.mapping.variable}}`,
+				`${rangeFrom ? `${search.mapping.variable}>=${rangeFrom}` : ''} ${rangeTo ? `${search.mapping.variable}<=${rangeTo}` : ''}`
+			)
+	);
 </script>
 
 <form
 	class="[&_label]:text-subtle mb-2 grid grid-cols-[1fr_1fr_auto] items-end gap-2 p-2"
 	action=""
-	on:submit={handleSubmit}
+	onsubmit={handleSubmit}
 >
 	<div class="flex flex-col gap-1">
-		<label class="sr-only" for="facet-range-from">{$page.data.t('general.from')}</label>
+		<label class="sr-only" for="facet-range-from">{page.data.t('general.from')}</label>
 		<input
 			class="bg-input h-8 rounded-sm border border-neutral-300 px-2 py-1"
 			id="facet-range-from"
@@ -43,12 +48,12 @@
 			min="1000"
 			max="2099"
 			step="1"
-			placeholder={`${$page.data.t('general.from')} (${$page.data.t('general.year')})`}
+			placeholder={`${page.data.t('general.from')} (${page.data.t('general.year')})`}
 			bind:value={rangeFrom}
 		/>
 	</div>
 	<div class="flex flex-col gap-1">
-		<label class="sr-only" for="facet-range-to">{$page.data.t('general.to')}</label>
+		<label class="sr-only" for="facet-range-to">{page.data.t('general.to')}</label>
 		<input
 			class="bg-input h-8 rounded-sm border border-neutral-300 px-2 py-1"
 			id="facet-range-to"
@@ -56,7 +61,7 @@
 			min="1000"
 			max="2099"
 			step="1"
-			placeholder={`${$page.data.t('general.to')} (${$page.data.t('general.year')})`}
+			placeholder={`${page.data.t('general.to')} (${page.data.t('general.year')})`}
 			bind:value={rangeTo}
 		/>
 	</div>
@@ -64,12 +69,12 @@
 		class="btn btn-primary size-8"
 		disabled={!rangeFrom && !rangeTo}
 		type="submit"
-		aria-label={$page.data.t('general.apply')}
+		aria-label={page.data.t('general.apply')}
 	>
 		<BiSearch class="text-base" />
 	</button>
 
-	{#each searchParams as [name, value]}
+	{#each searchParams as [name, value], index (index)}
 		{#if name !== '_q'}
 			<input type="hidden" {name} {value} />
 		{/if}
