@@ -1,16 +1,20 @@
 <script lang="ts">
 	import type { ItemDebugInfo } from '$lib/types/search';
-	import { page } from '$app/stores';
+	import { page } from '$app/state';
 
-	export let debugInfo: ItemDebugInfo;
+	interface Props {
+		debugInfo: ItemDebugInfo;
+	}
+
+	let { debugInfo }: Props = $props();
 	let score = debugInfo.score;
 
 	function fmt(x: number) {
-		return x.toLocaleString($page.data.locale, { maximumFractionDigits: 2 });
+		return x.toLocaleString(page.data.locale, { maximumFractionDigits: 2 });
 	}
 
 	function fmtPercent(x: number) {
-		return (x * 100).toLocaleString($page.data.locale, { maximumFractionDigits: 2 });
+		return (x * 100).toLocaleString(page.data.locale, { maximumFractionDigits: 2 });
 	}
 </script>
 
@@ -31,7 +35,7 @@
 			</tr>
 		</thead>
 		<tbody>
-			{#each score.perField as field}
+			{#each score.perField as field (field.name)}
 				<tr
 					title="haystack: {field.haystack.slice(0, 1)}{field.haystack.length > 1
 						? ` + ${field.haystack.length - 1}`
