@@ -1,6 +1,10 @@
 <script lang="ts">
 	import type { ItemDebugInfo } from '$lib/types/search';
-	export let debugInfo: ItemDebugInfo;
+	interface Props {
+		debugInfo: ItemDebugInfo;
+	}
+
+	let { debugInfo }: Props = $props();
 
 	let uniqueFields = debugInfo.score.perField.filter(
 		(field, index, array) => array.findIndex((f) => f.name === field.name) == index
@@ -16,7 +20,7 @@
 			</tr>
 		</thead>
 		<tbody>
-			{#each uniqueFields as field}
+			{#each uniqueFields as field (field.name)}
 				<tr
 					class="align-top"
 					title="haystack: {field.haystack.slice(0, 1)}{field.haystack.length > 1
@@ -25,7 +29,7 @@
 				>
 					<td>{field.name}</td>
 					<td>
-						{#each field.haystack as h}
+						{#each field.haystack as h, index (index)}
 							<p>{h}</p>
 						{/each}
 					</td>
