@@ -1,17 +1,26 @@
 <script lang="ts">
 	import getPageTitle from '$lib/utils/getPageTitle';
-	import SiteHeader from './SiteHeader.svelte';
 	import SiteFooter from './SiteFooter.svelte';
-	export let data;
+	import AppBar from './AppBar.svelte';
+	import { page } from '$app/state';
+	const { data, children } = $props();
+
+	const isFindRoute = $derived(page.route.id === '/(app)/[[lang=lang]]/find');
 </script>
 
 <svelte:head>
 	<title>{getPageTitle()}</title>
 	<base href={data.base} />
 </svelte:head>
-<SiteHeader />
-<main>
-	<slot />
-</main>
-<SiteFooter />
+<AppBar />
+{#if isFindRoute}
+	<div class="flex flex-1 flex-col">
+		{@render children()}
+	</div>
+{:else}
+	<main class="flex flex-1 flex-col">
+		{@render children()}
+	</main>
+	<SiteFooter />
+{/if}
 <div id="floating-elements-container"></div>
