@@ -2,22 +2,22 @@
 	import { page } from '$app/state';
 	import type { LocaleCode } from '$lib/i18n/locales';
 	import type { FacetGroup } from '$lib/types/search';
-	import { ShowLabelsOptions } from '$lib/types/decoratedData';
+	import { ExpandedState } from '$lib/types/userSettings';
 	import {
-		DEFAULT_FACETS_SHOWN,
+		CUSTOM_FACET_SORT,
 		DEFAULT_FACET_SORT,
-		CUSTOM_FACET_SORT
+		DEFAULT_FACET_VALUES_SHOWN,
+		MY_LIBRARIES_FILTER_ALIAS
 	} from '$lib/constants/facets';
 	import { getUserSettings } from '$lib/contexts/userSettings';
 	import { getMatomoTracker } from '$lib/contexts/matomo';
 	import { popover } from '$lib/actions/popover';
-	import FacetRange from './FacetRange.svelte';
-	import DecoratedData from '../DecoratedData.svelte';
+	import FacetValue from '$lib/components/find/FacetValue.svelte';
+	import FacetRange from '$lib/components/find/FacetRange.svelte';
 	import BiChevronRight from '~icons/bi/chevron-right';
 	import BiSortDown from '~icons/bi/sort-down';
-	import BiCheckSquareFill from '~icons/bi/check-square-fill';
-	import BiSquare from '~icons/bi/square';
 	import BiInfo from '~icons/bi/info-circle';
+	import BiPencil from '~icons/bi/pencil';
 
 	// Todo: Rename FacetGroup -> Facet (facets -> items/facetItems)
 
@@ -25,21 +25,30 @@
 		group: FacetGroup;
 		locale: LocaleCode;
 		searchPhrase: string;
+		isDefaultExpanded: boolean;
 	};
 
-	let { group, locale, searchPhrase }: FacetGroupProps = $props();
+	let { group, locale, searchPhrase, isDefaultExpanded }: FacetGroupProps = $props();
 
 	const matomoTracker = getMatomoTracker();
 	const userSettings = getUserSettings();
 
+	let detailsEl: HTMLDetailsElement | undefined = undefined;
+
 	const maxItems = group.maxItems;
 	const totalItems = group.facets.length;
-	let defaultItemsShown = $state(DEFAULT_FACETS_SHOWN);
+	let defaultItemsShown = $state(DEFAULT_FACET_VALUES_SHOWN);
 
 	let currentSort = $state(
 		userSettings.facetSort?.[group.dimension] ||
 			CUSTOM_FACET_SORT[group.dimension as keyof typeof CUSTOM_FACET_SORT] ||
 			DEFAULT_FACET_SORT
+	);
+
+	let isUserOrDefaultExpanded = $derived(
+		userSettings.facetExpanded?.[group.dimension]
+			? userSettings.facetExpanded?.[group.dimension] === ExpandedState.OPEN
+			: isDefaultExpanded
 	);
 
 	const sortOptions = [
@@ -64,10 +73,11 @@
 			}
 			if (currentSort === 'alpha.asc') {
 				return a.str.localeCompare(b.str, page.data.locale);
-			} else {
-				// hits.desc
-				return b.totalItems < a.totalItems ? -1 : 1;
 			}
+			if (a.totalItems === b.totalItems) {
+				return 0;
+			}
+			return b.totalItems < a.totalItems ? -1 : 1; // hits.desc
 		})
 	);
 
@@ -82,9 +92,10 @@
 
 	const shownItems = $derived(filteredItems.filter((facet, index) => index < defaultItemsShown));
 	let hasHits = $derived(filteredItems.length > 0);
+	let expanded = $derived(isUserOrDefaultExpanded || (searchPhrase && hasHits));
 	const canShowMoreItems = $derived(filteredItems.length > defaultItemsShown);
 	const canShowFewerItems = $derived(
-		!canShowMoreItems && filteredItems.length > DEFAULT_FACETS_SHOWN
+		!canShowMoreItems && filteredItems.length > DEFAULT_FACET_VALUES_SHOWN
 	);
 	const maxItemsReached = $derived(totalItems === maxItems);
 
@@ -97,6 +108,13 @@
 			$matomoTracker.trackEvent('Facet sort', group.dimension, target.value);
 		}
 	}
+
+	function saveUserExpanded(e: Event): void {
+		e.preventDefault();
+		if (detailsEl) {
+			userSettings.saveFacetExpanded(group.dimension, !detailsEl.open);
+		}
+	}
 </script>
 
 <li
@@ -104,10 +122,11 @@
 	class:has-hits={hasHits}
 	data-dimension={group.dimension}
 >
-	<details class="relative" open>
+	<details class="relative" open={!!expanded} bind:this={detailsEl}>
 		<summary
-			class="flex min-h-9 w-full cursor-pointer items-center gap-2 pr-8 text-xs font-medium"
+			class="hover:bg-primary-100 flex min-h-9 w-full cursor-pointer items-center gap-2 pr-12 pl-3 text-xs font-medium"
 			data-testid="facet-toggle"
+			onclick={saveUserExpanded}
 		>
 			<span class="arrow text-subtle transition-transform">
 				<BiChevronRight />
@@ -115,14 +134,12 @@
 			<span class="flex-1 whitespace-nowrap">{group.label}</span>
 		</summary>
 		<!-- sorting -->
-		<div
-			class="facet-sort btn btn-primary absolute top-0 right-0 m-1 size-6 border-0 p-0"
-			data-testid="facet-sort"
-		>
+		<div class="facet-sort absolute top-0 right-2 size-8" data-testid="facet-sort">
 			<select
+				name={group.dimension}
 				bind:value={currentSort}
 				onchange={saveUserSort}
-				class="size-full appearance-none text-transparent"
+				class="btn btn-primary size-full appearance-none border-0 text-transparent"
 				aria-label={page.data.t('sort.sort') + ' ' + page.data.t('search.filters')}
 			>
 				{#each sortOptions as option (option.value)}
@@ -130,7 +147,7 @@
 					>
 				{/each}
 			</select>
-			<BiSortDown class="pointer-events-none absolute top-0 right-0 m-1 text-base" />
+			<BiSortDown class="pointer-events-none absolute top-0 right-0 m-2 text-base" />
 		</div>
 		<div class="text-2xs">
 			{#if group.search && !(searchPhrase && hasHits)}
@@ -138,65 +155,48 @@
 				<FacetRange search={group.search} />
 			{/if}
 			<ol
-				class="border-l-neutral ml-1.5 flex max-h-72 flex-col overflow-x-clip overflow-y-auto border-l pr-0.5 pl-1.5 sm:max-h-[437px]"
+				class="flex max-h-72 flex-col overflow-x-clip overflow-y-auto sm:max-h-[453px]"
 				data-testid="facet-list"
 			>
 				{#each shownItems as facet (facet.view['@id'])}
-					<li>
-						<a
-							class="facet-link hover:bg-primary-50 flex items-end justify-between gap-2 p-1 pl-2 font-normal no-underline"
-							href={facet.view['@id']}
-						>
-							<span class="truncate" title={facet.str}>
-								{#if 'selected' in facet}
-									<!-- checkboxes -->
-									<span class="sr-only"
-										>{facet.selected ? page.data.t('search.activeFilter') : ''}</span
-									>
-									<div class="mr-1 inline-block text-xs" aria-hidden="true">
-										{#if facet.selected}
-											<BiCheckSquareFill class="text-accent" />
-										{:else}
-											<BiSquare class="text-subtle" />
-										{/if}
-									</div>
-								{/if}
-								<span>
-									<DecoratedData data={facet.object} showLabels={ShowLabelsOptions.Never} />
-									{#if facet.discriminator}
-										<span class="text-subtle">({facet.discriminator})</span>
-									{/if}
-								</span>
-							</span>
-							{#if facet.totalItems > 0}
-								<span class="badge" aria-label="{facet.totalItems} {page.data.t('search.hits')}"
-									>{facet.totalItems.toLocaleString(locale)}</span
-								>
-							{/if}
-						</a>
+					<li class="facet-group-list-value hover:bg-primary-100 flex">
+						<FacetValue {facet} {locale} />
+						{#if 'alias' in facet && facet.alias === MY_LIBRARIES_FILTER_ALIAS}
+							<a
+								href="/my-pages"
+								class="btn btn-primary mr-2 border-0"
+								aria-label={page.data.t('search.changeLibraries')}
+							>
+								<BiPencil />
+							</a>
+						{/if}
 					</li>
 				{/each}
 			</ol>
-			<div class="text-2xs flex items-center justify-between">
+			<div class="text-2xs flex flex-col justify-start">
 				<!-- 'show more' btn -->
 				{#if canShowMoreItems || canShowFewerItems}
 					<button
-						class="mt-2 ml-5 h-6"
+						class="hover:bg-primary-100 w-full"
 						onclick={() =>
 							canShowMoreItems
 								? (defaultItemsShown = totalItems)
-								: (defaultItemsShown = DEFAULT_FACETS_SHOWN)}
+								: (defaultItemsShown = DEFAULT_FACET_VALUES_SHOWN)}
 					>
-						{canShowMoreItems ? page.data.t('search.showMore') : page.data.t('search.showFewer')}...
+						<span class="ml-4.5 block border-l border-l-neutral-200 py-1.5 pr-3 pl-4 text-left">
+							{canShowMoreItems
+								? page.data.t('search.showMore')
+								: page.data.t('search.showFewer')}...
+						</span>
 					</button>
 				{/if}
 				<!-- limit reached info -->
 				{#if maxItemsReached && (canShowFewerItems || (!canShowMoreItems && searchPhrase))}
 					<button
-						class="text-error bg-severe-50 mt-2 flex items-center gap-1 rounded-sm px-2 py-1"
+						class="text-error bg-severe-50 m-3 flex items-center gap-1 rounded-sm px-2 py-1"
 						use:popover={{
 							title: page.data.t('facet.limitText'),
-							placeAsSibling: true
+							placeAsSibling: false
 						}}
 					>
 						<span>{page.data.t('facet.limitInfo')}</span>
@@ -210,19 +210,29 @@
 </li>
 
 <style lang="postcss">
-	@reference "../../../app.css";
-
 	details[open] {
 		& .arrow {
-			@apply rotate-90;
+			rotate: 90deg;
 		}
 		& .facet-sort {
-			@apply block;
+			display: block;
 		}
+
+		& summary:hover {
+			background-color: inherit;
+		}
+	}
+
+	.facet-group-list-value:has(.btn:hover) {
+		background-color: inherit;
 	}
 
 	/* hide sorting for bool filters */
 	li[data-dimension='boolFilters'] details[open] .facet-sort {
-		@apply hidden;
+		display: none;
+	}
+
+	li[data-dimension='accessFilters'] details[open] .facet-sort {
+		display: none;
 	}
 </style>
