@@ -9,18 +9,18 @@
 	};
 
 	const { data }: PaginationProps = $props();
-
-	const numberOfPages = 7;
-
 	const { first, last, next, previous, totalItems, itemsPerPage, itemOffset, maxItems } =
 		$derived(data);
+
+	const MAX_PAGES_SM = 3;
+	const MAX_PAGES_MD = 7;
+
 	const showPagination = $derived(data.items.length > 0 && totalItems > itemsPerPage);
 	const currentPage = $derived(Math.floor(itemOffset / itemsPerPage) + 1);
 	const lastItem = $derived(totalItems > maxItems ? maxItems : totalItems);
 	const lastPage = $derived(Math.ceil(lastItem / itemsPerPage));
-	const sequenceSize = $derived(numberOfPages > lastPage ? lastPage : numberOfPages);
 
-	const pageSequence = $derived.by(() => {
+	function getPages(sequenceSize: number) {
 		let pages = [];
 		let halfSequence = Math.floor(sequenceSize / 2);
 
@@ -46,7 +46,7 @@
 			pages.push({ page: i, offset: itemsPerPage * (i - 1) });
 		}
 		return pages;
-	});
+	}
 
 	function getOffsetLink(offset: number) {
 		let o = offset < 0 ? 0 : offset;
@@ -54,9 +54,37 @@
 	}
 </script>
 
+{#snippet sequence(maxSize: number)}
+	{@const pages = getPages(Math.min(maxSize, lastPage))}
+	{#if pages[0].page > 2}
+		<li class="text-2xs flex h-9 w-5 items-end justify-center pb-2 sm:w-9"><span>...</span></li>
+	{/if}
+	<!-- page sequence -->
+	{#each pages as p (p.offset)}
+		{#if p.page !== 1 && p.page !== lastPage}
+			<li>
+				<a
+					class={['btn btn-primary', p.page === currentPage ? 'bg-accent-50' : 'border-0']}
+					href={getOffsetLink(p.offset)}
+					aria-label="{page.data.t('search.page')} {p}"
+					aria-current={p.page === currentPage ? 'page' : null}
+					>{p.page.toLocaleString(page.data.locale)}</a
+				>
+			</li>
+		{/if}
+	{/each}
+	{#if lastPage - pages[pages.length - 1].page > 1}
+		<li class="text-2xs flex h-9 w-5 items-end justify-center pb-2 sm:w-9"><span>...</span></li>
+	{/if}
+{/snippet}
+
 {#if showPagination}
-	<nav aria-label={page.data.t('search.pagination')} data-testid="pagination">
-		<ul class="flex justify-center overflow-hidden p-4 sm:px-6">
+	<nav
+		aria-label={page.data.t('search.pagination')}
+		data-testid="pagination"
+		class="pagination mt-4 py-4"
+	>
+		<ol class="flex items-center justify-center">
 			<!-- prev -->
 			{#if previous}
 				<li>
@@ -76,29 +104,18 @@
 					href={first['@id']}>1</a
 				>
 			</li>
-			{#if pageSequence[0].page > 2}
-				<li class="hidden h-9 w-6 items-end pb-1 sm:flex"><span>...</span></li>
-			{/if}
-			<!-- page sequence -->
-			{#each pageSequence as p (p.offset)}
-				{#if p.page !== 1 && p.page !== lastPage}
-					<li>
-						<a
-							class={[
-								'btn btn-primary',
-								p.page === currentPage ? 'bg-accent-50' : 'hidden border-0 sm:flex'
-							]}
-							href={getOffsetLink(p.offset)}
-							aria-label="{page.data.t('search.page')} {p}"
-							aria-current={p.page === currentPage ? 'page' : null}
-							>{p.page.toLocaleString(page.data.locale)}</a
-						>
-					</li>
-				{/if}
-			{/each}
-			{#if lastPage - pageSequence[pageSequence.length - 1].page > 1}
-				<li class="hidden h-9 w-6 items-end pb-1 sm:flex"><span>...</span></li>
-			{/if}
+			<!-- sm sequence -->
+			<li class="flex sm:hidden">
+				<ol class="sequence flex items-center justify-center">
+					{@render sequence(MAX_PAGES_SM)}
+				</ol>
+			</li>
+			<!-- md sequence -->
+			<li class="hidden sm:flex">
+				<ol class="sequence flex items-center justify-center">
+					{@render sequence(MAX_PAGES_MD)}
+				</ol>
+			</li>
 			<!-- last -->
 			<li>
 				<a
@@ -118,18 +135,33 @@
 					>
 				</li>
 			{/if}
-		</ul>
+		</ol>
 	</nav>
 {/if}
 
-<style>
-	nav li > * {
-		margin-inline: calc(var(--spacing) * 0.5);
-	}
+<style lang="postcss">
+	.pagination {
+		@reference 'tailwindcss';
 
-	nav li > a {
-		min-height: calc(var(--spacing) * 9);
-		min-width: calc(var(--spacing) * 9);
-		padding-inline: calc(var(--spacing) * 2);
+		& li > a {
+			min-height: calc(var(--spacing) * 8);
+			min-width: calc(var(--spacing) * 8);
+		}
+
+		& .sequence {
+			margin-inline: calc(var(--spacing) * 0);
+		}
+
+		@variant md {
+			& li > * {
+				margin-inline: calc(var(--spacing) * 0.5);
+			}
+
+			& li > a {
+				min-height: calc(var(--spacing) * 9);
+				min-width: calc(var(--spacing) * 9);
+				padding-inline: calc(var(--spacing) * 2);
+			}
+		}
 	}
 </style>
