@@ -1,8 +1,7 @@
 <script lang="ts">
-	import { page } from '$app/stores';
-	import { goto } from '$app/navigation';
+	import { page } from '$app/state';
 	import { relativizeUrl } from '$lib/utils/http';
-	import type { QualifierSuggestion, SuperSearchResultItem } from '$lib/types/search';
+	import type { SuperSearchResultItem } from '$lib/types/search';
 	import DecoratedData from '$lib/components/DecoratedData.svelte';
 	import { ShowLabelsOptions } from '$lib/types/decoratedData';
 	import { LxlLens } from '$lib/types/display';
@@ -16,19 +15,11 @@
 		item: SuperSearchResultItem;
 		getCellId: (cellIndex: number) => string;
 		isFocusedCell: (cellIndex: number) => boolean;
-		addQualifier: (qualifier: QualifierSuggestion) => void;
 	};
 
-	const { item, getCellId, isFocusedCell, addQualifier }: Props = $props();
+	const { item, getCellId, isFocusedCell }: Props = $props();
 	const resourceId = $derived(relativizeUrl(item?.['@id']));
-
-	function handleClickPrimaryAction() {
-		if (item?.qualifiers.length) {
-			addQualifier(item?.qualifiers[0]);
-		} else if (resourceId) {
-			goto(resourceId);
-		}
-	}
+	const primaryAddQualifierLink = $derived(item.qualifiers?.[0]?._q || resourceId);
 </script>
 
 {#snippet resourceSnippet(item: SuperSearchResultItem)}
@@ -36,22 +27,22 @@
 		<span
 			class="text-subtle order-1 ml-auto hidden rounded-sm px-1.5 py-0.5 text-xs whitespace-nowrap sm:inline"
 		>
-			{$page.data.t('search.add')}
+			{page.data.t('search.add')}
 
 			<span class="hidden lowercase lg:inline">
 				{item.qualifiers[0].label}
 			</span>
 		</span>
 	{:else}
-		<div class="sr-only">{$page.data.t('search.goTo')}</div>
+		<div class="sr-only">{page.data.t('search.goTo')}</div>
 	{/if}
 	<div class="resource grid grid-cols-[40px_minmax(0,_1fr)] items-center gap-2">
 		<SuggestionImage {item} />
 		<div class="resource-content">
 			<hgroup
-				class="resource-heading flex gap-1 overflow-hidden text-xs font-medium whitespace-nowrap"
+				class="resource-heading grid gap-1 overflow-hidden text-xs font-medium whitespace-nowrap"
 			>
-				<h2 class="inline-block max-w-[40vw] truncate sm:max-w-[33vw]">
+				<h2 class="truncate">
 					<DecoratedData
 						data={item[LxlLens.CardHeading]}
 						showLabels={ShowLabelsOptions.Never}
@@ -60,8 +51,8 @@
 					/>
 				</h2>
 				{#if item[LxlLens.CardBody]?._display?.[0]}
-					<p class="inline-block truncate">
-						<span class="divider">{' • '}</span>
+					<p class="truncate">
+						<span class="divider">{' · '}</span>
 						<DecoratedData
 							data={item[LxlLens.CardBody]?._display[0]}
 							showLabels={ShowLabelsOptions.Never}
@@ -76,17 +67,17 @@
 					{item.typeStr}
 				</strong>
 				{#if item.typeStr?.length}
-					<span class="divider">{' • '}</span>
+					<span class="divider">{' · '}</span>
 				{/if}
-				{#each item?.[LensType.WebCardFooter]?._display as obj}
+				{#each item?.[LensType.WebCardFooter]?._display as obj, index (index)}
 					{#if 'hasInstance' in obj}
 						{@const instances = getInstanceData(obj.hasInstance)}
 						{#if instances?.years}
-							<span class="divider">{' • '}</span>
+							<span class="divider">{' · '}</span>
 							<span>
 								{#if instances.count > 1}
 									{instances?.count}
-									{$page.data.t('search.editions')}
+									{page.data.t('search.editions')}
 									{`(${instances.years})`}
 								{:else}
 									{instances.years}
@@ -109,14 +100,9 @@
 
 <div class="suggestion flex h-14 items-stretch" class:qualifier={item.qualifiers.length}>
 	{#if item.qualifiers.length}
-		<button
-			type="button"
-			id={getCellId(0)}
-			class:focused-cell={isFocusedCell(0)}
-			onclick={handleClickPrimaryAction}
-		>
+		<a href={primaryAddQualifierLink} id={getCellId(0)} class:focused-cell={isFocusedCell(0)}>
 			{@render resourceSnippet(item)}
-		</button>
+		</a>
 		<button
 			type="button"
 			class="more w-14 items-center justify-center p-0"
@@ -129,12 +115,12 @@
 					use:dropdownMenu={{
 						menuItems: [
 							...item.qualifiers.map((qualifier) => ({
-								label: `${$page.data.t('search.addAs')} ${qualifier.label.toLocaleLowerCase()}`,
-								action: () => addQualifier(qualifier)
+								label: `${page.data.t('search.addAs')} ${qualifier.label.toLocaleLowerCase()}`,
+								href: qualifier._q
 							})),
 							{
-								label: `${$page.data.t('search.goToResource')}`,
-								action: () => goto(resourceId as string)
+								label: `${page.data.t('search.goToResource')}`,
+								href: resourceId || ''
 							}
 						],
 						placeAsSibling: true
@@ -145,7 +131,7 @@
 			{/key}
 		</button>
 	{:else}
-		<a href={resourceId}>
+		<a href={resourceId} id={getCellId(0)}>
 			{@render resourceSnippet(item)}
 		</a>
 	{/if}
@@ -191,6 +177,8 @@
 	}
 
 	.resource-heading {
+		grid-template-columns: auto auto;
+
 		& :global(.transliteration) {
 			display: none;
 		}
@@ -201,7 +189,7 @@
 	}
 
 	.resource-footer {
-		/* hide dangling divider • */
+		/* hide dangling divider · */
 		& .divider {
 			display: none;
 		}
