@@ -1,7 +1,9 @@
-export const DEFAULT_FACETS_SHOWN = 5;
+export const DEFAULT_FACETS_EXPANDED = 4;
+export const DEFAULT_FACET_VALUES_SHOWN = 5;
 export const DEFAULT_FACET_SORT = 'hits.desc';
 export const CUSTOM_FACET_SORT = {
 	bibliography: 'alpha.asc',
 	itemHeldBy: 'alpha.asc'
 };
+export const ACCESS_FILTERS = ['freeOnline', 'alias-myLibraries'];
 export const MY_LIBRARIES_FILTER_ALIAS = 'alias-myLibraries';
