@@ -1,4 +1,4 @@
-import type { DisplayDecorated } from './xl';
+import type { DisplayDecorated, FramedData } from './xl';
 
 export type BibIdObj = {
 	bibId: string;
@@ -7,6 +7,7 @@ export type BibIdObj = {
 	onr: string | null;
 	isbn: string[];
 	issn: string[];
+	str: string;
 };
 
 export type HoldingsByInstanceId = {
@@ -28,4 +29,16 @@ export type DecoratedHolder = {
 	obj: DisplayDecorated;
 	sigel: string;
 	str: string;
+};
+
+export type FullHolderBySigel = {
+	[sigel: string]: FramedData;
+};
+
+export type ItemLinksForHolder = {
+	[sigel: string]: { [linkType: string]: string[] };
+};
+
+export type ItemLinksByBibId = {
+	[id: string]: ItemLinksForHolder;
 };
