@@ -14,11 +14,17 @@ function getInstanceData(instances: ResourceData) {
 
 		const res = jmespath.search(instances, query) as string[] | null;
 		if (res) {
-			years = res
+			const NUM_NEW = 3;
+			const NUM_OLD = 1;
+			const y = res
 				.filter((el, i, arr) => !isNaN(parseInt(el)) && arr.indexOf(el) === i)
 				.sort()
-				.filter((el, i, arr) => i === 0 || i === arr.length - 1)
-				.join('-');
+				.filter((el, i, arr) => i <= NUM_OLD || i >= arr.length - NUM_NEW);
+			if (y.length <= NUM_NEW + NUM_OLD) {
+				years = y.join(', ');
+			} else {
+				years = y.slice(0, NUM_OLD).join(', ') + ' … ' + y.slice(-NUM_NEW, y.length).join(', ');
+			}
 		}
 
 		return { count, years };
