@@ -8,23 +8,22 @@ function getLabelFromMappings(
 	pageMapping?: DisplayMapping[],
 	suggestMapping?: DisplayMapping[]
 ) {
-	const nonQuotedKey = key.replace(/^"(.*)"$/, '$1');
-	const nonQuotedValue = value && value.replace(/^"(.*)"$/, '$1');
 	const bestSuggestMapping = suggestMapping?.length ? suggestMapping : prevSuggestMapping;
 
-	const pageLabels = iterateMapping(nonQuotedKey, nonQuotedValue, pageMapping);
-	const suggestLabels = iterateMapping(nonQuotedKey, nonQuotedValue, bestSuggestMapping);
+	const pageLabels = iterateMapping(key, value, pageMapping);
+	const suggestLabels = iterateMapping(key, value, bestSuggestMapping);
 
 	const keyLabel = suggestLabels.keyLabel || pageLabels.keyLabel;
 	const valueLabel = suggestLabels.valueLabel || pageLabels.valueLabel;
 	const invalid = suggestLabels.invalid || pageLabels.invalid;
+	const removeLink = suggestLabels.removeLink || pageLabels.removeLink;
 
 	if (suggestMapping?.length) {
 		// save latest mapping as fallback for error responses etc
 		prevSuggestMapping = suggestMapping;
 	}
 
-	return { keyLabel, valueLabel, invalid };
+	return { keyLabel, valueLabel, removeLink, invalid };
 }
 
 function iterateMapping(
@@ -34,6 +33,7 @@ function iterateMapping(
 ) {
 	let keyLabel: string | undefined;
 	let valueLabel: string | undefined;
+	let removeLink: string | undefined;
 	let invalid: boolean = false;
 
 	if (mapping && Array.isArray(mapping)) {
@@ -53,15 +53,17 @@ function iterateMapping(
 					if (isLinked && value === el?._value && el?.displayStr) {
 						// only use atomic ranges for linked values
 						valueLabel = el.displayStr;
+						removeLink = el.up?.['@id'];
 					}
 				} else if (!key && value === el?._value && el?.displayStr) {
 					// ...unless a filter alias (no key, only value)
 					valueLabel = el.displayStr;
+					removeLink = el.up?.['@id'];
 				}
 			});
 		}
 	}
-	return { keyLabel, valueLabel, invalid };
+	return { keyLabel, valueLabel, removeLink, invalid };
 }
 
 export default getLabelFromMappings;
