@@ -1,11 +1,20 @@
 import { pushState } from '$app/navigation';
 import isFnurgel from '$lib/utils/isFnurgel';
-import type { BibIdObj, HoldersByType, HoldingsByInstanceId } from '$lib/types/holdings';
-import { LensType, type FramedData } from '$lib/types/xl';
+import type {
+	BibIdObj,
+	DecoratedHolder,
+	HoldersByType,
+	HoldingsByInstanceId,
+	ItemLinksByBibId,
+	ItemLinksForHolder
+} from '$lib/types/holdings';
+import { LensType, type FramedData, JsonLd, BibDb } from '$lib/types/xl';
 import type { LocaleCode } from '$lib/i18n/locales';
 import type { LibraryItem, UserSettings } from '$lib/types/userSettings';
-import { relativizeUrl } from '$lib/utils/http';
+import { relativizeUrl, stripAnchor } from '$lib/utils/http';
 import { DisplayUtil, toString } from '$lib/utils/xl.js';
+import getAtPath from '$lib/utils/getAtPath';
+import { holdersCache } from '$lib/utils/holdersCache.svelte';
 
 export function getHoldingsLink(url: URL, value: string) {
 	const newSearchParams = new URLSearchParams([...Array.from(url.searchParams.entries())]);
@@ -19,7 +28,7 @@ export function handleClickHoldings(
 	id: string
 ) {
 	event.preventDefault();
-	pushState(event.currentTarget.href, { ...state, holdings: id });
+	pushState(event.currentTarget.href, { ...state, holdings: stripAnchor(id) });
 }
 
 export function getSelectedHolding(value: string, instanceIdsByType: { [key: string]: string[] }) {
@@ -42,6 +51,20 @@ function sortHoldings(holdings) {
 		}
 		return 0;
 	});
+}
+
+export function getHoldersCount(mainEntity): number {
+	const instances = mainEntity['@reverse']?.instanceOf;
+	const holders = new Set();
+	instances?.forEach((instance) => {
+		const holdings = instance['@reverse']?.itemOf;
+		holdings?.forEach((h) => {
+			if (h.heldBy) {
+				holders.add(h.heldBy['@id']);
+			}
+		});
+	});
+	return holders.size;
 }
 
 export function getHoldingsByInstanceId(
@@ -72,13 +95,24 @@ export function getHoldingsByInstanceId(
 	}, {});
 }
 
-export function getBibIdsByInstanceId(mainEntity, record): Record<string, BibIdObj> {
-	return mainEntity['@reverse']?.instanceOf?.reduce((acc, instanceOfItem) => {
-		const id = relativizeUrl(instanceOfItem['@id'])?.replace('#it', '');
+export function getBibIdsByInstanceId(
+	mainEntity,
+	displayUtil: DisplayUtil,
+	record,
+	locale: LocaleCode
+): Record<string, BibIdObj> {
+	return mainEntity['@reverse']?.instanceOf?.reduce((acc, instance) => {
+		const id = relativizeUrl(instance['@id'])?.replace('#it', '');
 
-		const bibId = instanceOfItem.meta?.controlNumber || record?.controlNumber;
-		const type = instanceOfItem['@type'];
-		const holders = instanceOfItem['@reverse']?.itemOf?.map((i) => i?.heldBy?.sigel);
+		const bibId = instance.meta?.controlNumber || record?.controlNumber;
+		const type = instance['@type'];
+		const holders = instance['@reverse']?.itemOf?.map((i) => i?.heldBy?.sigel);
+		const publication = instance.publication;
+		let str = '';
+		if (publication) {
+			str =
+				toString(displayUtil.lensAndFormat(instance.publication[0], LensType.Token, locale)) || '';
+		}
 
 		// add Legacy Libris III system number for ONR param
 		let onr = null;
@@ -90,7 +124,7 @@ export function getBibIdsByInstanceId(mainEntity, record): Record<string, BibIdO
 
 		const isbn: string[] = [];
 		const issn: string[] = [];
-		instanceOfItem.identifiedBy?.forEach((el: { '@type': string; value: string }) => {
+		instance.identifiedBy?.forEach((el: { '@type': string; value: string }) => {
 			if (el['@type'] === 'ISBN') {
 				isbn.push(el.value);
 			}
@@ -111,7 +145,8 @@ export function getBibIdsByInstanceId(mainEntity, record): Record<string, BibIdO
 				holders,
 				onr,
 				isbn,
-				issn
+				issn,
+				str
 			}
 		};
 	}, {});
@@ -184,4 +219,147 @@ export function getMyLibsFromHoldings(
 		}
 	}
 	return Object.values(result);
+}
+
+export async function fetchHoldersIfAbsent(allHolders: DecoratedHolder[]) {
+	const cachedHolders = holdersCache.holders;
+	for (const h of allHolders) {
+		const id = h.obj?.['@id'];
+
+		if (h.sigel && cachedHolders && !cachedHolders[h.sigel]) {
+			const response = await fetch(`${id}?framed=true`, {
+				headers: { Accept: 'application/ld+json' }
+			});
+			if (response.ok) {
+				const resJson = await response.json();
+				const libraryMainEntity = resJson['mainEntity'] as FramedData;
+				if (libraryMainEntity) {
+					cachedHolders[h.sigel] = libraryMainEntity;
+				}
+			} else {
+				console.error(`Could not fetch holder data for ${id}`);
+			}
+		}
+	}
+}
+
+export function getItemLinksByBibId(
+	bibIdsByInstanceId: Record<string, BibIdObj>,
+	locale: LocaleCode,
+	displayUtil: DisplayUtil
+): ItemLinksByBibId {
+	const linksByInstanceId: ItemLinksByBibId = {};
+	for (const bibIdObj of Object.values(bibIdsByInstanceId || [])) {
+		const linksForHolder: ItemLinksForHolder = {};
+		bibIdObj.holders?.forEach((sigel) => {
+			if (holdersCache.holders) {
+				const fullHolderData = holdersCache.holders[sigel];
+
+				const ilsPaths = [
+					[BibDb.ils, BibDb.bibIdSearchUri],
+					[BibDb.ils, BibDb.isbnSearchUri],
+					[BibDb.ils, BibDb.issnSearchUri]
+				];
+
+				const lopacPaths = [[BibDb.lopac, BibDb.bibIdSearchUriByLang]];
+
+				let linksToItem = getLinksToItemFor(bibIdObj, fullHolderData, ilsPaths, locale);
+				const lopacLinksToItem = getLinksToItemFor(bibIdObj, fullHolderData, lopacPaths, locale);
+
+				const linkTemplateEod = getAtPath(fullHolderData, [BibDb.eodUri], []);
+				if (linkTemplateEod && linkTemplateEod.length) {
+					linksToItem = [linkTemplateEod.replace(/%BIB_*ID%/, bibIdObj.bibId), ...linksToItem];
+				}
+
+				//TODO: rename
+				const allLinks: { [linkType: string]: string[] } = {};
+
+				const itemStatusUri = getAtPath(fullHolderData, [BibDb.ils, BibDb.itemStatusUri], []);
+
+				if (itemStatusUri && itemStatusUri.length) {
+					allLinks[BibDb.ItemStatus] = [itemStatusUri];
+				}
+
+				const linksToCatalog: string[] = [];
+				const linkToCatalog = getAtPath(fullHolderData, [BibDb.ils, 'url'], undefined);
+				if (linkToCatalog && linkToCatalog.length) {
+					linksToCatalog.push(linkToCatalog);
+					allLinks[BibDb.LinksToCatalog] = linksToCatalog;
+				}
+
+				const linksToSite: string[] = [];
+				const linkToSite = getAtPath(fullHolderData, ['url', JsonLd.ID], undefined);
+				if (linkToSite && linkToSite.length) {
+					linksToSite.push(linkToSite);
+					allLinks[BibDb.LinksToSite] = linksToSite;
+				}
+
+				const openingHoursList: string[] = [];
+				const openingHours = getAtPath(fullHolderData, [BibDb.openingHours], undefined);
+				if (openingHours && openingHours !== '') {
+					openingHoursList.push(openingHours);
+					allLinks[BibDb.OpeningHours] = openingHoursList;
+				}
+
+				const addresses: string[] = [];
+				const address = getAtPath(fullHolderData, [BibDb.address, '*'], undefined);
+				const postalAddress = address.find((a) => a[JsonLd.TYPE] === BibDb.postalAddress);
+				const visitingAddress = address.find((a) => a[JsonLd.TYPE] === BibDb.visitingAddress);
+
+				if (address && address.length) {
+					addresses.push(
+						toString(displayUtil.lensAndFormat(visitingAddress, LensType.Card, locale)) || ''
+					);
+					addresses.push(
+						toString(displayUtil.lensAndFormat(postalAddress, LensType.Card, locale)) || ''
+					);
+					allLinks[BibDb.Address] = addresses;
+				}
+
+				if (linksToItem.length) {
+					allLinks[BibDb.LinksToItem] = linksToItem;
+				}
+
+				if (lopacLinksToItem.length) {
+					allLinks[BibDb.LoanReserveLink] = lopacLinksToItem;
+				}
+
+				if (Object.keys(allLinks).length) {
+					linksForHolder[sigel] = allLinks;
+				}
+			}
+		});
+		linksByInstanceId[bibIdObj.bibId] = linksForHolder;
+	}
+	return linksByInstanceId;
+}
+
+function getLinksToItemFor(
+	bibIdObj: BibIdObj,
+	fullHolderData: FramedData,
+	paths: string[][],
+	locale: LocaleCode
+): string[] {
+	let linksToItem: string[] = [];
+	for (const path of paths) {
+		const linkTemplate = getAtPath(fullHolderData, path, []);
+		if (linkTemplate && linkTemplate.length) {
+			if (path.includes(BibDb.bibIdSearchUriByLang) && bibIdObj.bibId !== '') {
+				const linkTemplateLocalized =
+					linkTemplate[locale] || linkTemplate['sv'] || Object.values(linkTemplate)[0];
+				linksToItem = [linkTemplateLocalized.replace(/%BIB_*ID%/g, bibIdObj.bibId), ...linksToItem];
+			}
+			if (path.includes(BibDb.bibIdSearchUri) && bibIdObj.bibId !== '') {
+				// forms in the wild %BIB_ID%, %BIBID%, more???
+				linksToItem = [linkTemplate.replace(/%BIB_*ID%/g, bibIdObj.bibId), ...linksToItem];
+			}
+			if (path.includes(BibDb.isbnSearchUri) && bibIdObj.isbn.length) {
+				linksToItem = [linkTemplate.replace(/%ISBN%/g, bibIdObj.isbn), ...linksToItem];
+			}
+			if (path.includes(BibDb.issnSearchUri) && bibIdObj.issn.length) {
+				linksToItem = [linkTemplate.replace(/%ISSN%/g, bibIdObj.issn), ...linksToItem];
+			}
+		}
+	}
+	return linksToItem;
 }
