@@ -1,5 +1,6 @@
 import { expect, test } from '@playwright/test';
 import AxeBuilder from '@axe-core/playwright';
+import { DEFAULT_FACETS_EXPANDED } from '$lib/constants/facets';
 
 test.beforeEach(async ({ page }) => {
 	await page.goto('/find?_q=f&_limit=20&_offset=0&_sort=&_i=f');
@@ -15,7 +16,7 @@ test('page displays the site header', async ({ page }) => {
 });
 
 test('page has a search input', async ({ page }) => {
-	await expect(page.getByTestId('main-search')).toBeVisible();
+	await expect(page.getByTestId('main-search').nth(0)).toBeVisible();
 });
 
 test('can change the language', async ({ page }) => {
@@ -57,10 +58,49 @@ test('expanded filters have no detectable a11y issues', async ({ page }) => {
 test('sorting the facet sets a cookie', async ({ page, context }) => {
 	const beforeCookies = await context.cookies();
 	expect(beforeCookies).toEqual([]);
-	await page.getByTestId('facet-sort').first().getByRole('combobox').selectOption('alpha.asc');
+	await page.getByTestId('facet-sort').nth(1).getByRole('combobox').selectOption('alpha.asc');
 	const afterCookies = await context.cookies();
 	expect(afterCookies[0].name).toEqual('userSettings');
-	expect(afterCookies[0].value).toEqual('{%22facetSort%22:{%22rdf:type%22:%22alpha.asc%22}}');
+	expect(afterCookies[0].value).toEqual(
+		'{%22leadingPane%22:{%22open%22:true}%2C%22facetSort%22:{%22rdf:type%22:%22alpha.asc%22}}'
+	);
+});
+
+test('facet opened/closed state is preserved', async ({ page, context }) => {
+	const beforeCookies = await context.cookies();
+	expect(beforeCookies).toEqual([]);
+
+	const firstClosed = DEFAULT_FACETS_EXPANDED;
+
+	await expect(page.getByTestId('facet-list').first()).toBeVisible();
+	await expect(page.getByTestId('facet-list').nth(firstClosed)).toBeHidden();
+
+	await page.getByTestId('facet-toggle').first().click();
+	await page.getByTestId('facet-toggle').nth(firstClosed).click();
+
+	await expect(page.getByTestId('facet-list').first()).toBeHidden();
+	await expect(page.getByTestId('facet-list').nth(firstClosed)).toBeVisible();
+
+	await page.goto('/find?_q=f&_limit=20&_offset=0&_sort=&_i=f');
+
+	await expect(page.getByTestId('facet-list').first()).toBeHidden();
+	await expect(page.getByTestId('facet-list').nth(firstClosed)).toBeVisible();
+});
+
+test('myLibraries filter is visible (when not active)', async ({ page }) => {
+	await expect(page.getByText('Avgränsa till mina bibliotek')).toBeVisible();
+});
+
+test('select myLibraries filter adds filter alias to url', async ({ page }) => {
+	await expect(page).not.toHaveURL(/alias-myLibraries/);
+	await page.getByText('Avgränsa till mina bibliotek').click();
+	await expect(page).toHaveURL(/alias-myLibraries/);
+});
+
+test('myLibraries without favourite libraries shows a message', async ({ page }) => {
+	await expect(page.getByTestId('my-libraries-warning')).not.toBeVisible();
+	await page.getByText('Avgränsa till mina bibliotek').click();
+	await expect(page.getByTestId('my-libraries-warning')).toBeVisible();
 });
 
 // Comment out test that fails in CI
