@@ -14,8 +14,7 @@ import {
 	type RangeValue
 } from '@codemirror/state';
 import { syntaxTree } from '@codemirror/language';
-import { mount } from 'svelte';
-import QualifierComponent from './QualifierComponent.svelte';
+import { mount, type Component } from 'svelte';
 import insertQuotes from './insertQuotes.js';
 import { messages } from '$lib/constants/messages.js';
 import insertSpaceAroundQualifier from './insertSpaceAroundQualifier.js';
@@ -26,6 +25,15 @@ export type Qualifier = {
 	operator: string;
 };
 
+export type QualifierWidgetProps = {
+	key: string;
+	keyLabel?: string;
+	operator: string;
+	value?: string;
+	valueLabel?: string;
+	removeLink?: string;
+};
+
 export type GetLabelFunction = (
 	key: string,
 	value?: string
@@ -33,20 +41,20 @@ export type GetLabelFunction = (
 	keyLabel?: string;
 	valueLabel?: string;
 	invalid?: boolean;
+	removeLink?: string;
 };
 
-export type RemoveQualifierFunction = (qualifier: string) => void;
+type QualifierWidgetComponent = Component<QualifierWidgetProps>;
 
 class QualifierWidget extends WidgetType {
 	constructor(
 		readonly key: string,
 		readonly keyLabel: string | undefined,
-		readonly keyType: string | undefined,
+		readonly operator: string,
 		readonly value: string | undefined,
 		readonly valueLabel: string | undefined,
-		readonly operator: string,
-		readonly operatorType: string | undefined,
-		readonly removeQualifierFn: RemoveQualifierFunction | undefined
+		readonly qualifierWidget: QualifierWidgetComponent,
+		readonly removeLink: string | undefined
 	) {
 		super();
 	}
@@ -56,22 +64,21 @@ class QualifierWidget extends WidgetType {
 			this.keyLabel === other.keyLabel &&
 			this.operator === other.operator &&
 			this.value === other.value &&
-			this.valueLabel === other.valueLabel
+			this.valueLabel === other.valueLabel &&
+			this.removeLink === other.removeLink
 		);
 	}
 	toDOM(): HTMLElement {
 		const container = document.createElement('span');
-		container.style.cssText = `position: relative; display:inline-flex`;
-		mount(QualifierComponent, {
+		container.style.cssText = `position: relative; display:inline-flex; margin: 0 1px;`;
+		mount(this.qualifierWidget, {
 			props: {
 				key: this.key,
 				keyLabel: this.keyLabel,
-				keyType: this.keyType,
+				operator: this.operator,
 				value: this.value,
 				valueLabel: this.valueLabel,
-				operator: this.operator,
-				operatorType: this.operatorType,
-				removeQualifierFn: this.removeQualifierFn
+				removeLink: this.removeLink
 			},
 			target: container
 		});
@@ -80,8 +87,8 @@ class QualifierWidget extends WidgetType {
 }
 
 function lxlQualifierPlugin(
-	getLabelFn?: GetLabelFunction,
-	removeQualifierFn?: RemoveQualifierFunction
+	qualifierWidget: QualifierWidgetComponent,
+	getLabelFn?: GetLabelFunction
 ) {
 	let atomicRangeSet: RangeSet<RangeValue> = RangeSet.empty;
 
@@ -98,16 +105,14 @@ function lxlQualifierPlugin(
 					if (node.name === 'Qualifier') {
 						const keyNode = node.node.getChild('QualifierKey');
 						const key = keyNode ? doc.slice(keyNode?.from, keyNode?.to) : '';
-						const keyType = keyNode?.firstChild?.type.name;
 
 						const operatorNode = node.node.getChild('QualifierOperator');
 						const operator = operatorNode ? doc.slice(operatorNode?.from, operatorNode?.to) : '';
-						const operatorType = operatorNode?.firstChild?.type.name;
 
 						const valueNode = node.node.getChild('QualifierValue');
 						const value = valueNode ? doc.slice(valueNode?.from, valueNode?.to) : undefined;
 
-						const { keyLabel, valueLabel, invalid } = getLabelFn?.(key, value) || {};
+						const { keyLabel, valueLabel, removeLink, invalid } = getLabelFn?.(key, value) || {};
 
 						// Add qualifier widget
 						if (keyLabel || valueLabel) {
@@ -115,12 +120,11 @@ function lxlQualifierPlugin(
 								widget: new QualifierWidget(
 									key,
 									keyLabel,
-									keyType,
+									operator,
 									value,
 									valueLabel,
-									operator,
-									operatorType,
-									removeQualifierFn
+									qualifierWidget,
+									removeLink
 								)
 							});
 							const decorationRangeFrom = node.from;
@@ -144,6 +148,7 @@ function lxlQualifierPlugin(
 			});
 		}
 		atomicRangeSet = ranges.finish();
+
 		return Decoration.set(widgets, true);
 	}
 
