@@ -1,17 +1,19 @@
 <script lang="ts">
 	import SuperSearch from '$lib/components/SuperSearch.svelte';
-	import { lxlQualifierPlugin } from '$lib/index.js';
 	import { lxlQuery } from 'codemirror-lang-lxlquery';
 	import type { JSONValue } from '$lib/types/json.js';
 	import type { MockQueryResponse } from './api/find/+server.js';
 	import clearIconSvg from './icon-clear.svg';
 	import backIconSvg from './icon-arrow-left.svg';
+	import type { ExpandedContentParams } from '$lib/components/SuperSearch.svelte';
 
 	let isLoading: boolean | undefined = $state();
 	let hasData: boolean | undefined = $state();
 	let value = $state('');
 	let placeholder = $state('Search');
 	let useFormAttribute = $state(false);
+	let useCustomExpandedContent = $state(false);
+	let useWrappingArrowKeyNavigation = $state(false);
 
 	function handlePaginationQuery(searchParams: URLSearchParams, prevData: JSONValue) {
 		const paginatedSearchParams = new URLSearchParams(Array.from(searchParams.entries()));
@@ -36,6 +38,42 @@
 	}
 </script>
 
+{#snippet expandedContent({
+	resultsCount,
+	resultsSnippet,
+	getCellId,
+	isFocusedRow,
+	isFocusedCell
+}: ExpandedContentParams)}
+	<nav>
+		<div role="rowgroup">
+			{#each { length: 2 }, index}
+				{@const rowIndex = index + 1}
+				<div
+					role="row"
+					class="persistent-item"
+					class:focused={isFocusedRow(rowIndex)}
+					data-testid="persistent-item"
+				>
+					<button
+						type="button"
+						role="gridcell"
+						id={getCellId(rowIndex, 0)}
+						class:focused-cell={isFocusedCell(rowIndex, 0)}
+					>
+						Persistent item {rowIndex}
+					</button>
+				</div>
+			{/each}
+		</div>
+		{#if resultsCount}
+			<div role="rowgroup">
+				{@render resultsSnippet({ rowOffset: 3 })}
+			</div>
+		{/if}
+	</nav>
+{/snippet}
+
 <form action="test1">
 	<fieldset>
 		<legend>Supersearch component</legend>
@@ -54,31 +92,14 @@
 			paginationQueryFn={handlePaginationQuery}
 			transformFn={handleTransform}
 			language={lxlQuery}
-			extensions={[lxlQualifierPlugin()]}
 			toggleWithKeyboardShortcut
 			defaultInputCol={-1}
-			defaultResultRow={1}
+			defaultResultRow={0}
 			defaultResultCol={0}
 			form={useFormAttribute ? 'form-outside' : undefined}
+			expandedContent={useCustomExpandedContent ? expandedContent : undefined}
+			wrappingArrowKeyNavigation={useWrappingArrowKeyNavigation}
 		>
-			{#snippet startContent({ getCellId, isFocusedCell, isFocusedRow })}
-				<div role="rowgroup">
-					<div>Header for start items</div>
-					{#each { length: 3 } as _item, index}
-						{@const rowIndex = index + 1}
-						<div role="row" class="start-item" class:focused={isFocusedRow(rowIndex)}>
-							<button
-								type="button"
-								role="gridcell"
-								id={getCellId(rowIndex, 0)}
-								class:focused-cell={isFocusedCell(rowIndex, 0)}
-							>
-								Start item {rowIndex}{_item}
-							</button>
-						</div>
-					{/each}
-				</div>
-			{/snippet}
 			{#snippet inputRow({
 				expanded,
 				inputField,
@@ -128,16 +149,6 @@
 				>
 					Search
 				</button>
-			{/snippet}
-			{#snippet persistentResultItemRow({ getCellId, isFocusedCell })}
-				<div class="persistent-item" data-testid="persistent-item">
-					<a
-						href={`/test1#${getCellId(0)}`}
-						role="gridcell"
-						id={getCellId(0)}
-						class:focused-cell={isFocusedCell(0)}>Show all results</a
-					>
-				</div>
 			{/snippet}
 			{#snippet resultItemRow({ resultItem, getCellId, isFocusedCell, rowIndex })}
 				<div class="result-item" data-testid="result-item">
@@ -192,6 +203,22 @@
 		><input type="checkbox" bind:checked={useFormAttribute} data-testid="use-form-attribute" />
 		Use form attribute
 	</label>
+	<label
+		><input
+			type="checkbox"
+			bind:checked={useCustomExpandedContent}
+			data-testid="use-custom-expanded-content"
+		/>
+		Use custom expanded content
+	</label>
+	<label
+		><input
+			type="checkbox"
+			bind:checked={useWrappingArrowKeyNavigation}
+			data-testid="use-wrapping-arrow-key-navigation"
+		/>
+		Use wrapping arrow key navigation
+	</label>
 </fieldset>
 
 <style>
@@ -244,20 +271,7 @@
 		padding-left: 44px;
 	}
 
-	.persistent-item {
-		display: flex;
-		min-width: 480px;
-
-		& a {
-			display: flex;
-			flex: 1;
-			align-items: center;
-			min-height: 44px;
-			text-align: left;
-		}
-	}
-
-	.start-item,
+	.persistent-item,
 	.result-item {
 		display: flex;
 		align-items: flex-start;
